requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
colorama==0.4.6
discord.py==2.7.1
distro==1.9.0
docstring_parser==0.17.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.2
ruff==0.15.6
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
